+++ docker/server/scripts/deploy/staff.sh
@@ -10,22 +10,26 @@ BUILD_DIR=.next-build npm run build -- --only -vvv --filter=@kenthackenough/$PRO
 
 cd $REPO/projects/$PROJECT
 
-# Move old dist folder
-mv .next .next-old
+if [ -d ".next-build" ]; then
 
-# Copy over the new one
-if mv .next-build .next; then
-#   Start the project
-    if pm2 show $PROJECT; then
-        echo "pm2: start $PROJECT"
-        NODE_ENV=production pm2 restart $PROJECT --update-env
+    # Move old dist folder
+    mv .next .next-old
+
+    # Copy over the new one
+    if mv .next-build .next; then
+    #   Start the project
+        if pm2 show $PROJECT; then
+            echo "pm2: start $PROJECT"
+            NODE_ENV=production pm2 restart $PROJECT --update-env
+        else
+            echo "pm2: initialize $PROJECT"
+            pm2 start --name $PROJECT "npm run start -- --only --filter=@kenthackenough/$PROJECT"
+        fi
+        pm2 show $PROJECT
+        rm -rf .next-old
     else
-        echo "pm2: initialize $PROJECT"
-        pm2 start --name $PROJECT "npm run start -- --only --filter=@kenthackenough/$PROJECT"
+        mv .next-old .next
+        exit 1
     fi
-    pm2 show $PROJECT
-    rm -rf .next-old
-else
-    mv .next-old .next
-    exit 1
+
 fi
